chore: convert PluginsInitializer
