--- packages/core/admin/admin/src/components/PluginsInitializer/init.js
@@ -1,11 +1,0 @@
-const init = (plugins) => {
-  return {
-    plugins: Object.keys(plugins).reduce((acc, current) => {
-      acc[current] = { ...plugins[current] };
-
-      return acc;
-    }, {}),
-  };
-};
-
-export default init;
