+++ packages/core/helper-plugin/src/features/StrapiApp.tsx
@@ -22,7 +22,7 @@ interface MenuItem {
 interface Plugin {
   apis: Record<string, unknown>;
   injectionZones: Record<string, unknown>;
-  initializer: ({ setPlugin }: { setPlugin: (pluginId: string) => void }) => null;
+  initializer: React.ComponentType<{ setPlugin(pluginId: string): void }>;
   getInjectedComponents: (
     containerName: string,
     blockName: string
@@ -55,7 +55,7 @@ type RunHookWaterfall = <InitialValue, Store>(
   store: Store
 ) => unknown | Promise<unknown>;
 
-interface StrapiAppContextValue {
+export interface StrapiAppContextValue {
   menu: MenuItem[];
   plugins: Record<string, Plugin>;
   settings: Record<string, StrapiAppSetting>;
